Fix reset brightness
It turns out the lamp sets itself to brightness 2 when you
use BRUP to turn it on.

That's ... kinda strange, really, but, hey.
Fixed the bug this creates.

The only question:
Is it better to turn the brightness back down to 1...
Or leave it like this...

I'm going to leave it like this.