Added individual study lamps config
I tried adding individual codes to the study lamps but
realized that unless I sync them up to the main controller,
which means writing a much more complex controller, it just
makes everything *super* weird.

So I'm going to leave the code in place but the configuration
disabled.

Also, I'm a little concerned I screwed up the ability of the
lights to reliably understand specific commands from the Pi.
Whoops.

Also, adaptive lighting is *the worst thing in the world*.
Never turn it on.