Smart lamp
Well, now it's an actual smartlamp

Sort of.  LOL.

It has feedback.  The brightness feedback is off and there
seem to be various other issues.  But you did it.  You have
smart lamps with integrated remotes.  Kind of.

Sometimes homekit loses interest in the messages...  Or
something does - it could be the plugin losing interest in
the mqtt messages.  Not sure.